+++ hypervisor/include/arch/x86/mmu.h
@@ -326,8 +326,8 @@ int modify_mem(struct map_params *map_params, void *paddr, void *vaddr,
 int modify_mem_mt(struct map_params *map_params, void *paddr, void *vaddr,
 		       uint64_t size, uint32_t flags);
 int check_vmx_mmu_cap(void);
-int allocate_vpid(void);
-void flush_vpid_single(int vpid);
+uint16_t allocate_vpid(void);
+void flush_vpid_single(uint16_t vpid);
 void flush_vpid_global(void);
 void invept(struct vcpu *vcpu);
 bool check_continuous_hpa(struct vm *vm, uint64_t gpa, uint64_t size);
